requirements.txt
# License Server - Dependencies
# FastAPI & Server
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
gunicorn>=21.2.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.23
alembic>=1.12.1
asyncpg>=0.29.0
aiosqlite>=0.19.0

# Security & Cryptography
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.1
cryptography>=41.0.7
pynacl>=1.5.0

# Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Async & HTTP
httpx>=0.25.1
aiofiles>=23.2.1

# Utils
python-dateutil>=2.8.2
uuid6>=2023.5.2

# Dev
pytest>=7.4.3
pytest-asyncio>=0.21.1
black>=23.11.0
ruff>=0.1.6
